convert to midi

+++ cfp.py
@@ -45,6 +45,7 @@ import numpy as np
 np.seterr(divide='ignore', invalid='ignore')
 import scipy
 import scipy.signal
+import scipy.fftpack
 
 
 
@@ -52,7 +53,7 @@ def STFT(x, fr, fs, Hop, h):
     t = np.arange(Hop, np.ceil(len(x)/float(Hop))*Hop, Hop)
     N = int(fs/float(fr))
     window_size = len(h)
-    f = fs*np.linspace(0, 0.5, np.round(N/2), endpoint=True)
+    f = fs*np.linspace(0, 0.5, int(np.round(N/2)), endpoint=True)
     Lh = int(np.floor(float(window_size-1) / 2))
     tfr = np.zeros((int(N), len(t)), dtype=np.float)
